printf

+++ aux.c
@@ -23,7 +23,7 @@ int process_str(const char *str)
  */
 long int process_int(va_list args)
 {
-	long int num = va_arg(args, long int);
+	int num = va_arg(args, int);
 	int count = 0;
 
 	if (num < 0)
